refactor: 상품 상세 페이지 서버 컴포넌트로 변경
- SEO를 위해 서버 컴포넌트로 변경하고, 클라이언트 로직이 필요한 부분은 따로 컴포넌트로 분리함
